requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
python-dotenv
requests
pytest
